+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Improvements:
 * [OE-100](https://openlmis.atlassian.net/browse/OE-100): Added quantity-unit-toggle to Stock on Hand page
 * [OE-101](https://openlmis.atlassian.net/browse/OE-101): Added quantity-unit-toggle to Stock Card page
 * [OE-97](https://openlmis.atlassian.net/browse/OE-97): Added quantity-unit-toggle to Issue, Receive and Adjustments pages
+* [OE-108](https://openlmis.atlassian.net/browse/OE-108): Added quantity-unit-input to Issue, Receive and Adjustments pages, code refactor after adding quantityUnitCalculateService service
 
 2.1.8 / 2025-03-31
 ==================