+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Improvements:
 * [OE-101](https://openlmis.atlassian.net/browse/OE-101): Added quantity-unit-toggle to Stock Card page
 * [OE-97](https://openlmis.atlassian.net/browse/OE-97): Added quantity-unit-toggle to Issue, Receive and Adjustments pages
 * [OE-108](https://openlmis.atlassian.net/browse/OE-108): Added quantity-unit-input to Issue, Receive and Adjustments pages, code refactor after adding quantityUnitCalculateService service
+* [OLMIS-8163](https://openlmis.atlassian.net/browse/OLMIS-8163): Included user details in stock card. 
 
 2.1.8 / 2025-03-31
 ==================